+++ 0x06-pointers_arrays_strings/main.h
@@ -11,7 +11,7 @@ char *string_toupper(char *str);
 char *cap_string(char *str);
 char *leet(char *str);
 char *rot13(char *str);
-void print_number(int n);
+void print_number(long n);
 char *infinite_add(char *n1, char *n2, char *r, int size_r);
 void print_buffer(char *b, int size);
 int _putchar(char c);
